+++ e2e/__tests__/jsonReporter.test.ts
@@ -28,6 +28,8 @@ describe('JSON Reporter', () => {
     runJest('json-reporter', ['--json', `--outputFile=${outputFileName}`]);
     const testOutput = fs.readFileSync(outputFilePath, 'utf8');
 
+    expect(testOutput.endsWith('\n')).toBe(true);
+
     try {
       jsonResult = JSON.parse(testOutput);
     } catch (err: any) {
@@ -71,11 +73,15 @@ describe('JSON Reporter', () => {
   });
 
   it('outputs coverage report', () => {
-    const result = runJest('json-reporter', ['--json']);
+    const result = runJest('json-reporter', ['--json'], {
+      keepTrailingNewline: true,
+    });
     let jsonResult: FormattedTestResults;
 
     expect(result.stderr).toMatch(/1 failed, 1 skipped, 2 passed/);
     expect(result.exitCode).toBe(1);
+
+    expect(result.stdout.endsWith('\n')).toBe(true);
 
     try {
       jsonResult = JSON.parse(result.stdout);