+++ packages/expect-utils/src/types.ts
@@ -6,4 +6,15 @@
  *
  */
 
-export type Tester = (a: any, b: any) => boolean | undefined;
+import type {EqualsFunction} from './jasmineUtils';
+
+export type Tester = (
+  this: TesterContext,
+  a: any,
+  b: any,
+  customTesters: Array<Tester>,
+) => boolean | undefined;
+
+export interface TesterContext {
+  equals: EqualsFunction;
+}