+++ packages/jest-types/src/Config.ts
@@ -448,6 +448,7 @@ export type ProjectConfig = {
   moduleNameMapper: Array<[string, string]>;
   modulePathIgnorePatterns: Array<string>;
   modulePaths?: Array<string>;
+  preset?: string;
   prettierPath: string;
   resetMocks: boolean;
   resetModules: boolean;