+++ packages/jest-worker/src/workers/NodeThreadsWorker.ts
@@ -170,6 +170,9 @@ export default class ExperimentalWorker
   }
 
   private _onMessage(response: ParentMessage) {
+    // Ignore messages not intended for us
+    if (!Array.isArray(response)) return;
+
     let error;
 
     switch (response[0]) {
@@ -227,7 +230,8 @@ export default class ExperimentalWorker
         break;
 
       default:
-        throw new TypeError(`Unexpected response from worker: ${response[0]}`);
+        // Ignore messages not intended for us
+        break;
     }
   }
 
@@ -254,6 +258,18 @@ export default class ExperimentalWorker
         this._worker.postMessage(this._request);
       }
     } else {
+      // If the worker thread exits while a request is still pending, throw an
+      // error. This is unexpected and tests may not have run to completion.
+      const isRequestStillPending = !!this._request;
+      if (isRequestStillPending) {
+        this._onProcessEnd(
+          new Error(
+            'A Jest worker thread exited unexpectedly before finishing tests for an unknown reason. One of the ways this can happen is if process.exit() was called in testing code.',
+          ),
+          null,
+        );
+      }
+
       this._shutdown();
     }
   }